examples: seperate ws and rest examples
